LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-events.tentative-expected.txt

Popover

FAIL The "beforetoggle" event (listener) get properly dispatched for popovers assert_equals: expected 0 but got 1
FAIL The "beforetoggle" event (attribute) get properly dispatched for popovers assert_false: expected false got true
FAIL The "beforetoggle" event is cancelable for the "opening" transition assert_false: expected false got true
FAIL The "beforetoggle" event is not fired for element removal assert_false: expected false got true
FAIL The "toggle" event is coalesced assert_false: expected false got true
